socks work: move socks patch setup, http max_header adjustment to http module logging: print stack trace only if debugging
